+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/internal/MessageListItemDiffCallback.kt
@@ -77,6 +77,7 @@ internal object MessageListItemDiffCallback : DiffUtil.ItemCallback<MessageListI
             mentions = oldMessage.mentionedUsers != newMessage.mentionedUsers,
             footer = oldItem.showMessageFooter != newItem.showMessageFooter,
             poll = oldMessage.poll != newMessage.poll,
+            threadMode = oldItem.isThreadMode != newItem.isThreadMode,
         )
     }